refactor: separate configuration validation phases
